+++ eggs/Naaya/naaya/core/site_logging.py
@@ -1,4 +1,5 @@
 import os
+import re
 import logging
 
 from naaya.core.zope2util import get_zope_env, ofs_path
@@ -9,6 +10,13 @@ from Products.Naaya.interfaces import INySite
 log = logging.getLogger(__name__)
 
 SITES_LOG_PATH_VAR = 'SITES_LOG_PATH' # name of config env. var
+SUFFIX = '-site-logger' # for naming loggers
+
+ACCESS = 'ACCESS'
+USER_MAN = 'USER_MANAGEMENT'
+ALLOWED_SLUGS = {ACCESS: ("VIEW", "DOWNLOAD", ),
+                 USER_MAN: ("ASSIGNED", "UNASSIGNED", ),
+                }
 
 def get_site_logger(site):
     """
@@ -16,8 +24,8 @@ def get_site_logger(site):
 
     """
     site_logger_slug = ofs_path(site).strip('/').replace('/', ',')
-    logger = logging.getLogger('%s-logger' % site_logger_slug)
-    return (site_logger_slug, logger)
+    logger = logging.getLogger('%s%s' % (site_logger_slug, SUFFIX))
+    return logger
 
 def create_site_logger(site):
     """
@@ -25,33 +33,30 @@ def create_site_logger(site):
     Called on startup for all existing INySites or after creation of a site.
 
     """
-    site_logger_slug, logger = get_site_logger(site)
+    logger = get_site_logger(site)
+    logger.propagate = 0
+    logger.setLevel(logging.INFO)
+    custom_format = '%(asctime) %(message)'
     abs_path = get_zope_env(SITES_LOG_PATH_VAR)
     if abs_path:
         try:
             if not os.path.exists(abs_path):
                 os.makedirs(abs_path)
-            log_filename = os.path.join(abs_path, '%s.log' % site_logger_slug)
+            log_filename = os.path.join(abs_path, '%s.log' % logger.name)
             if not os.access(log_filename, os.F_OK):
                 f = open(log_filename, 'a').close()
             if not os.access(log_filename, os.W_OK):
                 log.warn(("Could not add file handler for site logger %r"
                            " (log file write permissions)"), site)
                 return logger
-            logger.propagate = 0
-
-            supported_keys = ('asctime', 'message', )
-
-            log_format = ' '.join(['%%({%d})' % index for index, key in enumerate(supported_keys)])
-            custom_format = log_format.format(*supported_keys)
-
             handler = logging.FileHandler(log_filename)
-            handler.setFormatter(JSONFormatter(custom_format))
-            logger.setLevel(logging.INFO)
-            logger.addHandler(handler)
         except Exception, e:
-            log.exception("Could not add file handler for site logger %r", site)
-
+            handler = logging.StreamHandler()
+            log.exception("Could not create file handler for site logger %r",
+                          site)
+        else:
+            handler.setFormatter(JSONFormatter(custom_format))
+            logger.addHandler(handler)
     return logger
 
 def init_site_loggers():
@@ -59,4 +64,46 @@ def init_site_loggers():
     import Zope2
     for ob in Zope2.app().objectValues():
         if INySite.providedBy(ob):
-            create_site_logger(ob)
+            try:
+                create_site_logger(ob)
+            except Exception, e:
+                log.exception("Exception creating site logger for %r", ob)
+
+
+## Common Site Logging Api ##
+
+def log_user_access(context, who, how):
+    """ On open/download of a content type """
+    if how not in ALLOWED_SLUGS[ACCESS]:
+        log.info("Invalid value for `how`: %r in logging access", how)
+        return
+    data = {
+        'type': ACCESS,
+        'action': how,
+        'user': who,
+        'content_path': ofs_path(context),
+    }
+    site_logger = get_site_logger(context.getSite())
+    site_logger.info(data)
+
+def log_user_management_action(context, who, whom, assigned, unassigned):
+    """
+    It is called where user management logging is required, guarantees
+    consistency in log messages.
+
+    * `who` and `whom` are user ids
+    * `assigned` and `unnasigned` are lists of roles, both can not be empty
+
+    """
+    site_logger = get_site_logger(context.getSite())
+    data = {
+        'type': USER_MAN,
+        'who': who,
+        'whom': whom,
+        'content_path': ofs_path(context),
+    }
+
+    if unassigned:
+        site_logger.info(dict(data, action='UNASSIGNED', roles=unassigned))
+    if assigned:
+        site_logger.info(dict(data, action='ASSIGNED', roles=assigned))