+++ eggs/Naaya/CHANGELOG.rst
@@ -1,5 +1,6 @@
 3.1.0 (unreleased)
 --------------------
+* feature: #988 logging user management actions [simiamih]
 * feature: #882 logging user actions in text files on disk [bogdatan]
 
 3.0.9 (2012-08-28)